generators/service-object-storage/templates/python/requirements.txt
python-swiftclient~=3.4.0
python-keystoneclient~=3.13.0